E2E test: create pyrefly suite infrstructure
